requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
tensorboard>=2.13.0

# Medical imaging
monai>=1.2.0
nibabel>=5.1.0
SimpleITK>=2.3.0
pydicom>=2.4.0
scikit-image>=0.21.0
opencv-python>=4.8.0
torchio>=0.19.0

# Data handling
pillow>=10.0.0
albumentations>=1.3.0

# Evaluation and metrics
seaborn>=0.12.0
scipy>=1.11.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Demo
streamlit>=1.25.0
gradio>=3.40.0

# Optional: Logging and tracking
wandb>=0.15.0
mlflow>=2.5.0

# Optional: Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
